add transactional

+++ src/main/java/com/msl/community/controller/PublishController.java
@@ -67,7 +67,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id")Integer id,
+    public String edit(@PathVariable(name = "id")Long id,
                        Model model){
 
         QuesetionDTO question = questionService.getById(id);
